requirements.txt
keyboard==0.13.1
pkg-resources==0.0.0
py-getch==1.0.1
